+++ HISTORY.md
@@ -1,8 +1,14 @@
 # jacman-core
-## [v1.0]
-* extracted from JacMan
+## [v1.2]
+* Hack for TEXAS bug
+* new config.ini, constants loaded in core.rb
 
 ## [v1.1]
 * 'Force gi' button
 * MODEL_DIR moved to Jacinthe/Tools/Template/Mail
+
+## [v1.0]
+* extracted from JacMan
+
+